requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
boto3>=1.28.0
python-dotenv>=1.0.0
streamlit-echarts>=0.4.0
Pillow>=10.0.0
tavily-python>=0.3.0
